+++ app/PostCardContent.tsx
