+++ pages/admin/index.tsx
@@ -4,8 +4,8 @@ import AdminMenu from '../../app/AdminMenu'
 import { useAppDispatch, useAppSelector } from '@store/config'
 import { adminVisitant } from '@actions/admin'
 import Select, { ActionMeta, SingleValue } from 'react-select';
-import VisitorInfo from './VisitorInfo';
-import RecentMember from './RecentMember'
+// import VisitorInfo from './VisitorInfo';
+// import RecentMember from './RecentMember'
 import RecentArticle from '../../app/RecentArticle'
 import Head from 'next/head'
 import { IRecentArticleContent, IRecentMemberContent } from '@typings/db'
